refactor granule deletion and update tests
